feat(pis-auth): move everything to single index file
